remove help header and move login component
